pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mediatransfer"
version = "0.1.0"
description = "Transfer media/assets between cloud providers while preserving metadata"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.34",
    "click>=8.1",
    "google-auth>=2.28",
    "google-auth-oauthlib>=1.2",
    "requests>=2.31",
    "Pillow>=10.2",
    "tqdm>=4.66",
    "python-dateutil>=2.9",
    "piexif>=1.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "moto[s3]>=5.0",
    "responses>=0.25",
]

[project.scripts]
mediatransfer = "mediatransfer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mediatransfer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
